Update README.md
Windows OS assistant
This is a windows application that looks around, responds to voice commands like "open firefox" and "exit" and plays various animations that give it a lifelike personality. It's very simple as it stands, like a few lines on a sheet of paper.

Included are the images that make up the animations, the python program, and a batch file that checks for dependencies and executes. All of this including the structure and means of executing may be changed if this isn't favorable.